moved delete logic; needs to reload on delete
